Add createUser endpoint
This commit adds a new endpoint to the UserController class.
The endpoint is a POST mapping for "/users" and allows clients to
create a new user. The user's password is hashed using the
PasswordUtils class before storing it in the database. The UserDao
class is responsible for creating the user in the database. The
createUser method returns a ResponseEntity with a success message and
HTTP status code 201 (CREATED) to indicate that the user was created
successfully.