Add Main class
This commit adds the Main class to the com.calebjones package. The Main
class is annotated with @SpringBootApplication, indicating it as the
entry point for the Spring Boot application. In the main() method,
there is a call to DBConnectionManager.clearDatabase() for testing
purposes, followed by SpringApplication.run() to start the
application.